fix : list position -greenhopper 방식 소수점 단위가 아니라 100단위로 계산되게 수정 -double -> BigDecimal

+++ src/main/java/com/example/panttegi/board/dto/BoardDetailListDto.java
@@ -3,6 +3,7 @@ package com.example.panttegi.board.dto;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Getter
@@ -11,6 +12,6 @@ public class BoardDetailListDto {
 
     private final Long id;
     private final String title;
-    private final Double position;
+    private final BigDecimal position;
     private final List<BoardDetailCardDto> cards;
 }
